feat: moved server context to explicit RouterContextProvider, which useLinguiServer now requires
- Reorganized src structure
- Use dependenciesMeta for better test project integration
- Fixed invalid import in the test project, closes #27
