+++ build.config.ts
@@ -2,13 +2,7 @@ import { defineBuildConfig } from "unbuild"
 
 export default defineBuildConfig({
   outDir: "dist",
-  entries: [
-    "src/index",
-    "src/lingui.client",
-    "src/lingui.server",
-    "src/routes",
-    "src/plugin/index",
-  ],
+  entries: ["src/index", "src/i18n.client", "src/i18n.server", "src/routes", "src/plugin/index"],
   externals: ["virtual:lingui-router-loader", "virtual:lingui-router-manifest"],
   declaration: true,
   rollup: {
