+++ src/components/LocaleLink.ts
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link, type LinkProps, type To } from "react-router"
-import { usePathLocale } from "./lingui"
+import { usePathLocale } from "../i18n"
 
 /**
  * A drop-in replacement for react-router's Link that automatically prefixes the
