+++ src/index.ts
@@ -1,6 +1,6 @@
+export * from "./components/I18nApp"
+export * from "./components/LocaleLink"
+export * from "./components/LocalePreload"
 export * from "./config"
-export * from "./I18nApp"
-export * from "./lingui"
-export * from "./LocaleLink"
-export * from "./LocalePreload"
+export * from "./i18n"
 export * from "./runtime"
