+++ test/app/locales/cs.po
@@ -25,7 +25,11 @@ msgstr "Nastala neočekávaná chyba."
 msgid "Continue"
 msgstr "Pokračovat"
 
-#: app/routes/hello.tsx:14
+#: app/routes/hello.tsx:8
+msgid "From loader too!"
+msgstr "Také z loaderu!"
+
+#: app/routes/hello.tsx:17
 msgid "Hello, World!"
 msgstr "Ahoj světe!"
