+++ test/app/routes/hello.tsx
@@ -1,19 +1,22 @@
+import { msg } from "@lingui/core/macro"
 import { Trans } from "@lingui/react/macro"
-import { data, useLoaderData } from "react-router"
+import { useLinguiServer } from "lingui-react-router/server"
+import { data, type LoaderFunctionArgs, useLoaderData } from "react-router"
 
-export function loader() {
-  return data({ msg: `From loader too!` })
+export function loader({ context }: LoaderFunctionArgs) {
+  const { _ } = useLinguiServer(context)
+  return data({ msg: _(msg`From loader too!`) })
 }
 
 export default function Hello() {
-  const loaderData = useLoaderData()
+  const loaderData = useLoaderData<typeof loader>()
 
   return (
     <div className="flex min-h-screen items-center justify-center">
       <h1 className="text-4xl font-bold">
         <Trans>Hello, World!</Trans>
       </h1>
-      <p>{loaderData?.msg}</p>
+      <p>{loaderData.msg}</p>
     </div>
   )
 }
