+++ test/package.json
@@ -15,10 +15,15 @@
     "@react-router/node": "^7.9.3",
     "@react-router/serve": "^7.9.3",
     "isbot": "^5.1.31",
-    "lingui-react-router": "link:..",
+    "lingui-react-router": "workspace:*",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
     "react-router": "^7.9.3"
+  },
+  "dependenciesMeta": {
+    "lingui-react-router": {
+      "inject": true
+    }
   },
   "devDependencies": {
     "@lingui/cli": "^5.5.0",
